Combine clear_crontab capistrano task into update_crontab
I ran into a problem when the Ruby version specified in the Gemfile was
changing. The clear_crontab task was running after the code was updated,
but before the bundle was installed. This led to an error due to the
discrepancy in the version of Ruby.

By combining the two tasks, none of the Whenever gem's code is executed
until after the bundle is installed.