Upgrade ruby-trello to actually fix rest-client vulnerabilities
The attempt at fixing the rest-client security vulnerabilities in the
previous commit were unsuccessful, because of a constraint on
rest-client by ruby-trello. By upgrading to the earliest version of
ruby-trello which supports the fixed version of rest-client we can
hopefully fix the vulnerability while reducing the chance of breaking
the (untested) code.