Use trello gem to print name of first board
Note that unfortunately the trello gem currently (rather unnecessarily)
has an indirect dependency on the unf gem which has native extensions.
When I was trying this previously, I thought I ran into problems because
the extensions were being built on my local machine using Mac OSX
(x86_64-darwin-17), but were being deployed and run in an AWS Lambda
function which is an x86_64-linux platform. Previously I had used Docker
to build the extensions for the correct platform, however, when I just
tried deploying and running the function as it stands (with Mac
OSX-build extensions), it seemed to work OK, so maybe the whole Docker
thing was unnecessary!