Write Trello board data to S3 bucket
I've chosen to use the board ID rather than a normalized version of the
board name for the S3 object key, because this is a lot less likely to
change over time.

There's no need to include a timestamp in the S3 object key, because the
S3 bucket has versioning enabled.