Remove alarm for errors from lambda functions
We've been seeing (typically 3 of) these alarms every time the backup
happens in recent weeks. The automatic retry mechanism for SNS has meant
that these problems have always resolved themselves. I considered
bumping up the threshold fo this alarm, but since we already have an
alarm on the dead letter queue which will be triggered when retries are
exhausted, I decided to remove this alarm altogether because it's always
likely to relate to temporary issues.

Note that we also have a separate check in healthchecks.io which ensures
that the backup has succeeded within 30 mins, so I don't think we're
increasing the risk of not knowing about backup failure by making this
change.