Signal failure to Healthchecks endpoint if any backup has failed
Trello: https://trello.com/c/8sJX4GLV

Previously the healthcheck would only fail if the checkBoardBackups
lambda function did not execute at least once every 24 hours. Even
though I was the one who added it, I've been misled by the way it was
working and wrongly though the backup was working when it wasn't.

That seems dangerous and so we now signal failure to the Healthchecks
endpoint if any of the backups fail. While this means there is some
overlap in the sending of the SNS email notification on backup failure
and the signalling of failure to Healthchecks, the former also gives
more detail on which board backups have failed and the latter gives us
confidence that the daily checkBoardBackups function continues to run.