Increase memory available to backupBoard lambda function 128MB -> 1024MB
Trello: https://trello.com/c/8sJX4GLV

The backup of the main "Go Free Range" Trello board has been repeatedly
failing in recent days and it looks as if the backupBoard lambda
function is running out of memory and being killed.

This isn't surprising, because we're downloading a single JSON document
which represents the whole board including archived card and this is
likely to continually grow over time. This will mean the memory
requirements of the function are also likely to grow over time.

Bumping the memory allowance might not be a long-term fix, but hopefully
it will do for now. The choice of 1024MB memory is completely arbitrary,
but hopefully it will be enough to keep things going for a while. I
don't think there's much danger we're going to break the bank if I've
increased the memory allowance more than necessary, because this
function only runs a few times a day and it doesn't take long to run.