Fix warning in use of find
I was seeing the following warning:

> find: warning: you have specified the global option -maxdepth after
the argument -type, but global options are not positional, i.e.,
-maxdepth affects tests specified before it as well as those specified
after it.  Please specify global options before other arguments.

James isn't seeing it on Mac OS so presumably it's due to a difference
in the `find` command between Mac and Linux. I'm hoping this doesn't
now cause a problem when running on a Mac.

In case it's useful, the version of find I'm using is:

$ find --version
find (GNU findutils) 4.8.0